Separated Homework1 Tasks into different projects
