+++ CoherentTasks.sln
@@ -3,7 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31911.196
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FirstSet", "FirstSet\FirstSet.csproj", "{6BDD7210-B9FC-48A5-8793-C6133F2A3986}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Task1.NumberListWith2TwosInTernary", "FirstSet\Task1.NumberListWith2TwosInTernary.csproj", "{6BDD7210-B9FC-48A5-8793-C6133F2A3986}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "HomeWork1", "HomeWork1", "{643AA6FC-3CD9-49FB-BD5A-8441CAF47699}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task2.ISBN.Number", "Task2\Task2.ISBN.Number.csproj", "{D6C3BF54-CC89-4215-B35F-23FCF33BEA40}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task3.SumOfListElements", "Task3\Task3.SumOfListElements.csproj", "{D43CF93F-5520-43CA-855C-B8CA082BDA14}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,9 +21,22 @@ Global
 		{6BDD7210-B9FC-48A5-8793-C6133F2A3986}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6BDD7210-B9FC-48A5-8793-C6133F2A3986}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6BDD7210-B9FC-48A5-8793-C6133F2A3986}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6C3BF54-CC89-4215-B35F-23FCF33BEA40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6C3BF54-CC89-4215-B35F-23FCF33BEA40}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6C3BF54-CC89-4215-B35F-23FCF33BEA40}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6C3BF54-CC89-4215-B35F-23FCF33BEA40}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D43CF93F-5520-43CA-855C-B8CA082BDA14}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D43CF93F-5520-43CA-855C-B8CA082BDA14}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D43CF93F-5520-43CA-855C-B8CA082BDA14}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D43CF93F-5520-43CA-855C-B8CA082BDA14}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{6BDD7210-B9FC-48A5-8793-C6133F2A3986} = {643AA6FC-3CD9-49FB-BD5A-8441CAF47699}
+		{D6C3BF54-CC89-4215-B35F-23FCF33BEA40} = {643AA6FC-3CD9-49FB-BD5A-8441CAF47699}
+		{D43CF93F-5520-43CA-855C-B8CA082BDA14} = {643AA6FC-3CD9-49FB-BD5A-8441CAF47699}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B0DD6C7D-EFC6-44B1-967E-0EBF104C2CE1}
