+++ FirstSet/Task1.NumberListWith2TwosInTernary.csproj
